+++ jackify-engine/Verbs/Encrypt.cs
@@ -4,6 +4,7 @@ using System.CommandLine.NamingConventionBinder;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Wabbajack.CLI.Builder;
+using Wabbajack.Common;
 using Wabbajack.Paths;
 using Wabbajack.Paths.IO;
 using Wabbajack.Services.OSIntegrated;
@@ -32,7 +33,7 @@ public class Encrypt
         var data = await input.ReadAllBytesAsync();
         _logger.LogInformation("Encrypting {bytes} bytes into `{key}`", data.Length, name);
         await data.AsEncryptedDataFile(name.ToRelativePath()
-            .RelativeTo(KnownFolders.WabbajackAppLocal.Combine("encrypted")));
+            .RelativeTo(JackifyConfig.GetDataDirectory().Combine("encrypted")));
         return 0;
     }
 }